requirements.txt
google-genai == 1.41.0
requests == 2.32.5
python-dotenv